Change AddressTagger to LocationTagger
