+++ requirements.txt
@@ -1,7 +1,8 @@
 Django==4.2
-geopy==2.3.0
+django_environ==0.10.0
 gpsphoto==2.2.3
 numpy==1.24.2
+Requests==2.28.2
 transformers==4.28.1
 exifread==3.0.0
 piexif==1.1.3
